refactor: split src/app.rs into src/app/ module
